requirements.txt
# Django Framework
Django==5.1.3

# PostgreSQL Database Adapter
psycopg2-binary==2.9.10

# API Framework
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0

# Testing Framework
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-xdist==3.6.1  # Parallel test execution

# Test Data Generation
factory-boy==3.3.1
Faker==33.1.0

# Code Quality
coverage==7.6.9
